Move injection of attributes from command line to `libsyntax_ext`
